Edit Utility class
