--- app/src/main/java/io/github/ovso/righttoknow/common/NetworkCheck.java
@@ -1,19 +1,0 @@
-package io.github.ovso.righttoknow.common;
-
-import android.content.Context;
-import android.net.ConnectivityManager;
-import android.net.NetworkInfo;
-
-/**
- * Created by jaeho on 2017. 8. 14
- */
-
-public class NetworkCheck {
-
-  public static boolean isOnline(Context context) {
-    ConnectivityManager connMgr =
-        (ConnectivityManager) context.getSystemService(Context.CONNECTIVITY_SERVICE);
-    NetworkInfo networkInfo = connMgr.getActiveNetworkInfo();
-    return (networkInfo != null && networkInfo.isConnected());
-  }
-}
