+++ app/src/main/java/io/github/ovso/righttoknow/common/Utility.java
@@ -1,0 +1,32 @@
+package io.github.ovso.righttoknow.common;
+
+import android.app.Activity;
+import android.content.Context;
+import android.net.ConnectivityManager;
+import android.net.NetworkInfo;
+import android.os.Build;
+import android.view.Window;
+import android.view.WindowManager;
+
+/**
+ * Created by jaeho on 2017. 8. 14
+ */
+
+public class Utility {
+
+  public static boolean isOnline(Context context) {
+    ConnectivityManager connMgr =
+        (ConnectivityManager) context.getSystemService(Context.CONNECTIVITY_SERVICE);
+    NetworkInfo networkInfo = connMgr.getActiveNetworkInfo();
+    return (networkInfo != null && networkInfo.isConnected());
+  }
+
+  public static void setStatusBarColor(Activity activity, int color) {
+    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
+      Window window = activity.getWindow();
+      window.addFlags(WindowManager.LayoutParams.FLAG_DRAWS_SYSTEM_BAR_BACKGROUNDS);
+      window.clearFlags(WindowManager.LayoutParams.FLAG_TRANSLUCENT_STATUS);
+      window.setStatusBarColor(color);
+    }
+  }
+}
